Refactor console domain actions to exist in separate files
This means that we're not storing everything in one file, otherwise it
quickly becomes unwieldy
