+++ core/src/main/java/google/registry/tools/EppToolCommand.java
@@ -136,7 +136,7 @@ abstract class EppToolCommand extends ConfirmingCommand implements CommandWithCo
       params.put("dryRun", dryRun);
       params.put("clientId", command.clientId);
       params.put("superuser", superuser);
-      params.put("xml", URLEncoder.encode(command.xml, UTF_8.toString()));
+      params.put("xml", URLEncoder.encode(command.xml, UTF_8));
       String requestBody =
           Joiner.on('&').withKeyValueSeparator("=").join(filterValues(params, Objects::nonNull));
       responses.add(
