+++ core/src/test/resources/google/registry/module/frontend/frontend_routing.txt
