+++ core/src/test/resources/google/registry/module/routing.txt
